Возвращает ветку master к начальному виду
